+++ source/extensions/omni.isaac.collision_sphere_editor/docs/CHANGELOG.md
@@ -1,12 +1,22 @@
 # Changelog
 
+## [0.2.0] - 2022-09-02
+
+### Changed
+
+- Enhanced collision sphere interpolation feature to generate spheres with
+  radii following a geometric sequence, positioned so as to produce a smooth
+  conical frustum in the limit of infinite spheres.
+
 ## [0.1.1] - 2022-09-02
 
 ### Removed
+
 - Removed unused dependencies
 - Removed from load function: automatic bookmark to file path outside of Isaac Sim directories.
 
 ## [0.1.0] - 2022-08-15
 
 ### Added
+
 - Initial version of Collision Sphere Editor Extension